pyproject.toml
[tool.poetry]
name = "airflow-cross-instance-datasets"
version = "0.1.0"
description = ""
authors = ["Julian de Ruiter <julianderuiter@gmail.com>"]
readme = "README.md"
packages = [{ include = "airbridge", from = "src" }]

[tool.poetry.scripts]
airbridge = 'airbridge.app:cli'

# [tool.poetry.plugins."airflow.plugins"]
# airbridge = "airbridge.plugin:AirbridgePlugin"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = "2.10.1"

# Workaround for Poetry issue: https://github.com/python-poetry/poetry/issues/9293.
pika = "^1.3.2"
typer = "^0.12.5"
aio-pika = "^9.4.3"
httpx = "^0.27.2"
pydantic-settings = "^2.5.2"
structlog = "^24.4.0"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"


[tool.poetry.group.dev.dependencies]
ipython = "^8.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
